server/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
Flask-SocketIO==5.5.1
qrcode==8.2
Pillow==11.3.0
PyJWT==2.10.1
python-dotenv==1.1.1
Werkzeug==3.1.3
